Uploaded first draft of tools and sample project
The methods that work are all in one file with some comments indicating
further work.  It is not compatible with the existing directory
structure (just wanted to get everything uploaded to get started).  Also
includes a sample project structure to be used for simple testing
purposes (the output filenames are correct but no anonymizing has
occurred in this version).